+++ app/models/taxon_name_relationship/original_combination/original_subform.rb
@@ -31,7 +31,7 @@ class TaxonNameRelationship::OriginalCombination::OriginalSubform < TaxonNameRel
     true
   end
 
-  def monomial_prefix
+  def monominal_prefix
     'subf.'
   end